chore: just moved Door and DoorsList component
